Run Flux tests before Jetpack/WordPress unit tests
When you run more than one test target via a test plan, the Xcode
console will show the logs for only the last target to run.

By running the Flux tests first and the Jetpack/WordPress last, we keep
what are most likely to be the logs we are interested into in the
console at the end of each test run.

It's useful to keep an eye on what's printed in the console when the
tests run because one can notice all sorts of runtime issues through
them.