Fixes #20245
This is a workaround to prevent the crash happening in #20245. The crash happens when the app attempts to clean up Quick Start tours. The clean up is done when the app determines that Jetpack features are no longer to be shown. However, the clean up produces a Core Data crash which I haven't understood properly yet. So a workaround is to simply not do the clean up and leave the Quick Start tours in Core Data. They won't be visible to the user.