Annotate a variable used only in `NSAssert` with `__unused`
CI failed to archive the build with:

```
WordPress/Classes/Services/AccountService.m:485:20: unused variable 'defaultAccount' [-Werror,-Wunused-variable]

WPAccount *defaultAccount = [WPAccount lookupDefaultWordPressComAccountInContext:context];
     ^
```

See
https://buildkite.com/automattic/wordpress-ios/builds/12837#0186b56a-1a05-422a-9ecc-1a0c0ce4cdb0
The error can be reproduced locally.

Evidently, our release configurations meet the condition described in
the `NSAssert` documentation:

> Assertions are disabled if the preprocessor macro NS_BLOCK_ASSERTIONS is defined or the ENABLE_NS_ASSERTIONS Xcode build setting is disabled.

See https://developer.apple.com/documentation/foundation/nsassert

An alternative approach would have been to modify the release
configurations, but this is a fix in the context of a release branch so
I wanted it to be as lightweight as possible.

Yet options were to silence the Clang error or to wrap the `NSAssert` in
a `#ifdef` precompiler check. The `__unused` annotation option seemed to
me the cleaner of them all.