Move release notes entry for #20275 to version 21.8.1
The fix in #20275 was originally made on top of the `release/21.9`
branch but we decided to ship it as a hotfix.

I could have amended 9dfe3379997faab2bab14698101959b5c3d814d1 with this
change, but decided to keep it in one piece to preserve a record of the
original change.