Use `replacePersistentStore` over `migratePersistentStore`
There have been performance issues on accounts which have a large number of
blogs. On the account which had the issue, there was a memory warning. This is likely due to
how `migratePersistentStore` works. The developer documentation on these two functions
isn't great but according to a Core Data document, `migratePersistentStore` does the following:

1. Creates a temporary persistence stack.
2. Mounts the old and new stores.
3. Loads all objects from the old store.
4. Migrates the objects to the new store.
5. The objects are given temporary IDs, then assigned to the new store. The new store then saves the newly assigned objects (committing them to the external repository).
6. Informs other stacks that the object IDs have changed (from the old to the new stores), which keeps the stack running after a migration.
7. Unmounts the old store.
8. Returns the new store.

Ref: https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/CoreData/PersistentStoreFeatures.html#//apple_ref/doc/uid/TP40001075-CH23-SW1

Even though the function name `replacePersistentStore` implies that it is replacing
something, it can be used to backup a database too. If the destination file does not
exist, it will create it.

According to an Apple engineer, `replacePersistentStore` will attempt to do an
APFS clone where possible. This greatly improves both the performance of the backup
and the memory footprint.

Quote:
> Additionally you should almost never use NSPersistentStoreCoordinator's migratePersistentStore... method but instead use the newer replacePersistentStoreAtURL.. (you can replace emptiness to make a copy). The former loads the store into memory so you can do fairly radical things like write it out as a different store type. It pre-dates iOS. The latter will perform an APFS clone where possible.

Ref: https://developer.apple.com/forums/thread/651325