Remove no-longer-required `brew update` from CI commands
We added this back in the day to address a Bintray issue that prevented
Homebrew from installing dependencies. See
3a1c00c7bd25483b0f0e6940bd449a01793ca441 and
https://github.com/wordpress-mobile/WordPress-iOS/pull/16373

The fix hasn't been required for a long time but the operation never
gave us any trouble until some weeks ago, see
78fcc6f3ce6c419de9c3a375c66a4ade96c8906f.

Now it's a good time to go back and cleanup.