Enable Core Data concurrency debug option in UI tests
The app crashes during UI tests when a Core Data concurrency issue occurs.

A Core Data concurrency issue is typically where a managed object is
accessed outside of the context (or the queue managed by the context).
Currently there are still many such issues in the app, post editing
being the most impacted one, which is why the new
`crashOnCoreDataConcurrencyIssue` option is disabled in `EditorGutenbergTests`.

But, I think it's still valuable to ensure that those areas that don't
have any Core Data concurrency issues to be maintained. And we should
expose the new issue if one does occur. Doing that in UI tests is one
easy way I can think of, which doesn't have a "blocking" factor—we can
always disable the check for specific test(s).