Remove Jetpack-only entry and WordPress-only label for WP 22.1 notes
Notice that the entry that was removed did not have the "Jetpack-only"
label but was still a Jetpack-only feature. I verified this by following
the instructions in
https://github.com/wordpress-mobile/WordPress-iOS/pull/20369
and not seeing the button to personalize the home screen. This is
consistent with only Jetpack offering a home screen that is enhanced
with cards.