Remove direct access to Blog after fetching in processSchedule
Refactor `addLocalNotification` and `addStaticNotifications` to
operate with value types instead of dealing with `Blog` directly.

There are still loose property reads to `Blog` within the method,
but `processSchedule` is private and its callers are guaranteeing
that the method is called from the Blog's context queue.

Further improvements will be made in the next iteration.