Configure table view
The init parameter is now converted to a local `Connection` model,
since the `PrepublishingAutoSharingModel` doesn't include the service
information within its connections.

Furthermore, the cells containing disabled connections are now disabled
once the remaining share limit is used up. This is to prevent the number
of enabled connections to go over the remaining share limit.