Sync blog instead of just the sharing limit
The view controller now stores `BlogService` instead of `JetpackSocialService`
to perform `syncBlog` when checkout is dismissed.

Additionally, the property `isSharingLimitReached` is updated to
`canInteractWithDisabledConnections`, and its value is now initialized
in the `init` method according to the sharing limit. This takes care
of an edge case where the remaining shares is 0, where all accounts
should've been disabled.

Lastly, added some docs and minor code organization.