Add delegate method for the social accounts sheet
The `PrepublishingViewController` conforms to the delegate protocol,
and performs a Core Data operation to persist the changes to the
connections and sharing message.

To optimize, the delegate method is only called once when the social
accounts sheet is dismissed. Changes to the connection states  are
stored in `connectionChanges`, and the delegate method will only be
called when there are changes in either the connections or the sharing
message.