Centralize commands to config Sim in UI tests in aggregate target
By moving them in an aggregate target on which the UI tests targets
depend, we DRY the setup as well as making the existence of this
capability more visible (in the targets list instead of buried in the
build phases).

Having the commands in a separate script file also makes it easier to
inspect diffs and working on them, given that the Xcode build phase
script editor window is far from a good text interface for coding.