Fall back to a default bottom inset value and apply a minimum content height
Previously, the `defaultBottomInset` was meant for popover views
on iPad since the device will return nil `safeAreaInsets`.

Now, the `defaultBottomInset` will be used for devices with zero-
valued safeAreaInsets (e.g., iPhone SE). This fixes the case where
the button may appear cropped because the drawer doesn't provide
enough "height" for the content.

Furthermore, for visual purposes, the content height is set to a
minimum of 300. This solves the case where users with unlimited
sharing and with only 1 connection could have the drawer positioned
too low, which could very easily be dismissed by mistake.