Clarify placeholders `stats.insights.totalLikes.guideText.singular`
One translator removed the placeholder for the second parameter (to be
substituted for the numeral one) which our linter picked up. Removing
placeholders will result in runtime crashes.

The idea with this `comment` update is to make it clearer that the
placeholder should remain, just in case the translation is picked up by
a new translator unfamiliar with the rule.