Replace hardcoded ASC team id with env var read for consistency
This is done for consistency. We've long established that it's safe to
have the ASC team id in clear on the repo. I decided to replace it with
an env var access so that, when we'll finally get 'round to removing
that moving part, we won't leave that one usage behind.