Fix no connections social dashboard card appearing
When a user sets up a connection on a screen that is pushed
onto the home dashboard, the no connections card does not
disappear. This is because the cards do not get reloaded with
new data.

To fix this, I added an observer to watch for the main context
changing. When that happens, if there is an update to the `blog`
object, the display state of the card gets updated.