Fix Zendesk identity creation flow in chat
When the user creates a Zendesk ticket via the chatbot, the app asks the user for their identity (email and name) if they haven't set it yet.
This flow wasn't working in the scenario where these weren't set. The fix involves passing in the view controller to use to present the identity form (an alert) on top of.