Manually fix `ro` string that didn't pass linting
This is a replay of 8b451114a3d364a34569285972b2d8301e9e19b2.
Evidently, the entry in GlotPress has yet to be approved despite it
being there for almost two weeks...

Comment from 8b451114a3d364a34569285972b2d8301e9e19b2 below:

I suggested a fix in GlotPress and tried to approve it using the admin
account, but got an error.

I'll follow up this internally, but in the meantime I fixed the
translation manually to complete the beta deployment.

https://translate.wordpress.org/projects/apps/ios/dev/ro/default/?filters%5Boriginal_id%5D=16177110&filters%5Bstatus%5D=either&filters%5Btranslation_id%5D=107327707