Update a `precondition` check to be a `guard`
When the user's auth token is invalid, the editor service will
attempt to initialize a `WordPressOrgRestApi` class. Due to there
still being an account with the blog, the `precondition` would
cause a crash.

Since the initialize and `apiBase` function are already nullable,
we can switch this `precondition` to a `guard` and then return `nil`.