Hide keyboard after sending chat bot message to reveal all messages
This is a soft workaround for mobile Safari issue that doesn't resize viewport when keyboard appears, causing all the content to shift up. It results in top messages hiding and double scroll appearing. This fix does not solve the whole issue but limits the chances of users not seeing the full content after asking a question