Better support for dynamic type and attributed text
If we set font on attributed text, we need to listen to notifications to automatically update dynamic type, otherwise the updated font will only appear after reopening the view.

To avoid listening to notifications in UITableViewCell, making most common attributed text cases rely on the font set directly on the label, and not on attributed text. This way dynamic type updates work out of the box.