Track loaded widgets when application is opened
Tracking all the loaded widgets when application is opened allows to see the percentage of unique daily users who use widgets and proportion of usage of each widget.

Given we only need an approximate tracking to understand the usage of widgets we don't need to rely on unpredictable widget loading events.

A single event also make it more likely to understand and track.