Add separator border view
Adds a slightly larger container view for the site icon image view
in order to have a separating border between the avatar image and
the icon image.

It's done this way so that we can have two separate borders. There's
one which gets added to the site icon image view and the second is
added to the new containing view.