Fallback to `false` in `ENV.fetch` for flag to bypass confirmation
`false` is the desired default value for the `bypass` parameter.

I noticed this when I replied no to the prompt but it did not abort as I
expected. See how the lane moved to `setbranchprotection` even though
the input was `n`

```
[19:13:19]: Ready to push changes to remote to let the automation configure it on GitHub? (y/n)
n
[19:14:13]: Aborting code completion. See you later.
[19:14:13]: ---------------------------------
[19:14:13]: --- Step: setbranchprotection ---
[19:14:13]: ---------------------------------
==========================================
This action (setbranchprotection) is deprecated
```

However, `false` vs `nil` is not the cause of the defective behavior.
That is due to the branch that's meant to abort only printing a message
about it and not calling `next`.