Rename the WordPressTest target to UnitTests
This is so that it will not look odd once the unit tests run on the
Jetpack scheme with the Jetpack app.

Notice that this also required renaming the target in the `Podfile`.

One additional note: I used `PRODUCT_NAME` as a proxy for the folder
name to define the path of `Info.plist` and other files in the build
settings. Out of the box, Xcode simply duplicates the paths, but in our
setup I think this gives us the flexibility to change the target name in
the future more easily.

Fun fact: `PrepublishingHeaderViewTests.swift` needed additional work
after the rename because, unlike the rest of the test files, it was
referenced via absolute path in the project.  ¯\_(ツ)_/¯