Move test for `DeepLinkSource` + `WidgetUrlSource` to WordPressTests
`DeepLinkSource` belongs in the WordPress target, at least for the
moment until we might move it to a dedicated underlying one.

By testing its interop with `WidgetUrlSource` at the WordPressTests
level we guarantee we still have that logic checked without having to
jump through hoops to get `DeepLinkSource` in the widgets target.