Move business logic for `WidgetDataReader` in package so we can test
Up to now, `WidgetDataReader` had to be part of the WordPress target for
the sole purpose of testing it.

We can't move `WidgetDataReader` in the package just yet because it
depends on `SelectSiteIntent` and `WPAppGroupName` which are entwined in
the rest of the app structure. But we can move all the "pure" business
logic there, enabling us to test that in isolation and at the package
level and, more importantly, remove the file from the WordPress target.

As a reminder, removing it from WordPress will make it possible to
switch the host app for the unit tests from Jetpack to WordPress.

We _could_ have simply added the file to the Jetpack target, but that
would have been perpetrating the inappropriate setup, instead of properly
addressing via better modularity.