Remove old checkout option for Gutenberg in `Podfile.lock`
This kept attempting to download the whole gutenberg-mobile with
submodules source code, which is one of the things the XCFramework
migration wanted to prevent.

My guess is that it was a setting `pod` wrote in the lockfile and that
has no way of being automatically removed, pending maybe upon removal of
the pod itself?