Fix crash when accessing widget week summary data outside bounds
endIdex is the array’s “past the end” position—that is, the position one greater than the last valid subscript argument.

If summaryData.count is n (5), endIndex will be equal to n (5). If we pass index 4, index + 1 will be outside the boundaries of summaryData, because array with n=5, has max index of n-1.