Track `Podfile.lock` change after `pod update Gutenberg`
Notice that the version apparently downgraded from 1.109.3 to 1.109.2.

The reason this happened is that we are on a merge branch from the
release branch into trunk.

The release branch commit from which this branch originated is
22b2216488e0576e8b9906aa72271d489175f34b.
The latest commit on trunk that was merged into here is
12065e659d39e59a7e9a01a45caf31112dee49b8.

At those commits, release/23.9 uses Gutenberg for Mobile Apps (GMA) v1.109.3 while trunk
v1.110.0-alpha2.

The versioning in GMA works in a way that only updates the version
value when a new stable version is shipped.
As such, v1.110.0-alpha2 results has having version 1.109.2 as can be
seen in the source code at
https://github.com/wordpress-mobile/gutenberg-mobile/blob/v1.110.0-alpha2/package.json#L3

Since the version of trunk that was merged in this branch uses
v1.110.0-alpha2, the computed version number tracked in `Podfile.lock`
is 1.109.2.