chore: Add constants for better site type tracking
We want to be able to track various site types with media events.

A `WPAppAnalyticsKeyIsJetpack` constant already exists, which stores an `is_jetpack` string. With this commit, we add to the list of constants to store `is_atomic` and `is_hosted_at_wpcom` strings.

This sets us up to reference those strings elsewhere in the codebase.