refactor: Remove redundant method
The tracking method that fires when a media upload errors out was refactored in the previous commit. As such, the original method for errors (which was originally added in https://github.com/wordpress-mobile/WordPress-iOS/pull/22328) is no longer used anywhere in the codebase and can be removed. It's removed with this commit.