Fix filtering by RSS in the subscription feed
When fetching the posts for RSS feeds, the API only returns 1 post
at a time. Due to this, it would be marked as having no more
additional posts.

To fix this, the `hasMore` boolean will return `true` for RSS feeds
if there are any posts that return from the API.