Define StatsTrafficModels that support DiffableDataSource in a single place
The diffable data source relies on both the identity and the equality of the items it manages. The identity is determined by the item's hash, and equality is determined by whether the item's content has changed. If the content of an item is considered to have changed (even if its hash hasn't), the diffable data source may decide to reload that item