Stop loading WP app asset on JP app
The JP app was loading an asset meant only for the WP app. This happens because the view controller containing this asset, JetpackFullscreenOverlayViewController, is loaded into memory before the code checks to confirm it's running on the WP app. The fix is to check first which app it's on, and then load the view controller if on the WP app.