Wait for build to finish processing in TestFlight so it distributes
It looks like the parameter to skip waiting for processing is in
conflict with the configuration to distribute the build to testers.

We can see this in action comparing the latest beta builds from
WordPress and WooCommerce.

- https://buildkite.com/automattic/wordpress-ios/builds/20777#018df168-f8f6-4a9c-8d19-041215a9e8d0
- https://buildkite.com/automattic/woocommerce-ios/builds/19923#018deafc-f600-434d-b0ca-c58b5ae658d6

The WordPress one logs:

```
[12:46:23]: Waiting for App Store Connect to finish processing the new build (24.3 - 24.3.0.4) for IOS
[12:46:24]: Successfully set the changelog for build
[12:46:24]: Using App Store Connect's default for notifying external testers (which is true) - set `notify_external_testers` for full control
```

While WooCommerce logs:

```
[06:55:42]: Waiting for App Store Connect to finish processing the new build (17.5 - 17.5.0.1) for IOS
[06:56:17]: Successfully finished processing the build 17.5 - 17.5.0.1 for IOS
[06:56:18]: Successfully set the changelog for build
[06:56:18]: Using App Store Connect's default for notifying external testers (which is true) - set `notify_external_testers` for full control
[06:56:20]: Distributing new build to testers: 17.5 - 17.5.0.1
```

This makes sense procedurally, because if Fastlane exits before the
build processed, then the build cannot be distributed. From a design
standpoint, it's confusing to be able to set those conflicting values.
(Rant over. Maybe I should make a PR in Fastlane instead of
complaining...)