Locally update the tag's `following` property after adding a tag
Previously, adding a tag from the Manage flow does not update the
tag's `following` property locally until the next time it's fetched.

Additionally, the model's `following` property seems to be affecting
the value of `showInMenu`. I've also added a convenience method that
ensures that `showInMenu` is also updated when `following` is updated.