Fix SiteStatsDetails not loading for Insight statSection that are not included to allInsights enum
allInsights only contains insights that are displayed in Stats Insights tab. There're other Insights types that are displayed in Stats Insights Details.

We shouldn't depend on statSection enum types on StatsDetails and make Insight or Period store calls if query exists. This way we ensure that such an issue doesn't repeat.