Fetch data for an insight that doesn't have any cache
Before the change we would load "error" status for a type that doesn't have cached values even though there could simply be no data for a given data type.

This change keeps the caching expiration mechanism but only if a given data type has cache. Otherwise, the data is fetched.