Do not upload a build to App Center during release cycle
We've been discussing internally the value of these builds and struggled
to make a strong case for them. The data confirmed this feeling, with
only a couple of downloads over the past months.

With the recent Microsoft announcement of [App Center shutting down next year](https://learn.microsoft.com/en-us/appcenter/retirement), the time seems ripe for removing this low value step.