Update ReaderFollowButton to accept ReaderDisplaySetting
The priority is still given to the `ButtonColor` object that's
passed through the parameter. The `ReaderDisplaySetting` only takes
precedence when there are no `ButtonColor` objects being passed.

The `@State` wrapper is removed from the `color` property since
it seems to be the cause of the button not updating colors after
the display setting is updated.