First pass: refactor WebCommentContentRenderer for custom styling
This allows the renderer to customize its style based on the
`ReaderDisplaySetting` object.

Some components need to be refactored in order to make everything
work; for example, the CSS file is located in the main bundle,
but we need the `loadHTMLString` to run on the WordPressShared's
bundle so that we can load the Noto font family.

To make this work, the `richCommentStyle.css` is loaded from main
bundle as a string and then injected to the HTML template.