Refactor font generation method for ReaderDisplaySetting
* Ensures that the custom weight is applied to the generated font.
* This also updates the default value of the `weight` parameter to
nil, to respect the inherent font weight from the `UITextStyle`.
* Lastly, but more importantly, this avoids the use of
`UIFontMetrics.scaledFont` because it's a bit prone to error; as in,
fonts cannot go through the `scaledFont` method twice; otherwise, it
will raise an exception. --Of course, this change also ensures that
the generated font scales along with the preferred content size
category.