Fix scrolling to the comment from notifications
This workaround does not stop navigating to the highlighted comment after the first unsuccessful attempt. Instead of that, when the comments pages are loaded one by one until the highlighted comment is loaded, we scroll down to the last available comment. It looks a bit laggy but reaches the destination after all.