Use an SVG (and manually apply blur)
Xcode’s SVG handling is interesting – it doesn’t properly handle filters, so the blur isn’t baked into the resulting image.

This is actually a feature, not a bug – the old encoding approach was worst-case scenario for image size – a partially-transparent image with gradients (from the blur) and noise applied overtop; and because Xcode turns the SVG into a PNG for us it would get unexpected large very quickly.

Doing the blur manually takes just as much CPU time as it would if we were processing the SVG within iOS, but results in a much smaller image inside the bundle.