Open AppStore directly rather than through SKStoreProductViewController
SKStoreProductViewController is being opened in different flows, some of which are opened on separate UIWindows. SKStoreProductViewController has strict rules of presentation, one of which is it has to be presented in a separate modal view. To avoid issues such as these, I'm reverting to a safe option of opening AppStore directly