Add the FAB to the Reader.
Unlike in My Sites, the button is added to the `window` instead of
the current view. This prevents the button from being obstructed by
the ReaderStreamViewController.

This ensures that the button is at the "right" Z-index: it's always
shown on top of the feed and doesn't animate when switching streams.
Also, the sheet is correctly displayed on top of the FAB.