Don't delete previous site topic
There's a cascade delete behavior between ReaderSiteTopic and its
`posts`. When we stack the same site stream several times, the
previous stream's posts will get deleted and raise an exception
because all the managed objects have turned to faults.

We can't also rely on the `inUse` mechanism (i.e., don't delete
if inUse == true), because the stream view controller sets the
topic's `inUse` property to false on deinit--which leads to the
same issue as above.

Note that removing this may reintroduce the bug in #22885, but
I guess not crashing is a better choice for now.