Register table view cells as early as possible
This prevents loading unregistered cells (specific to the Discover
stream), due to a `layoutIfNeeded` call from the superclass
`ReaderStreamViewController`, in the `refreshTableViewHeaderLayout`
method.

Previously, in the `ReaderCardsStreamViewController`, the cell
registration was called *after* `super.viewDidLoad`. This opens up
the possibility for the superclass to trigger `layoutIfNeeded` and
caused the cards stream to crash.