Remove unused `release_management_in_ci.rb` lanes
We no longer need repo-level automation to trigger release lanes. We
have a dedicated internal app to do so. Keeping the lanes here would be
more of a liability in terms of code that never gets used and is bound
to become out-of-date.

In the rare occasion where one might need to trigger those builds
manually, it can be done via the Buildkite GUI or API directly.

To verify they are unused I run `git grep -e 'trigger_.*_in_ci'`, which
returned only the lane declarations. Of course, this does not guarantee
there are no references outside of the repo, but based on my
understanding of our Automation I think it's safe to assume there are
none. (But just to be sure, I grepped our backend codebase for `_in_ci`,
and got no results.)

See also conversation at
https://github.com/Automattic/simplenote-ios/pull/1655#discussion_r1753900531