fix: Discard initial selection for single select
Workaround existing logic that dismisses the dialog if `allowMuliple` is
true and selection is not empty. The existing logic allows for quickly
selecting a single media item, but inhibits the ability to showcase a
pre-existing selection for single-select contexts. This diverges from
the web experience, but is the easiest path forward to avoid a
significant refactor of the logic that would impact all other existing
use cases--Gutenberg Mobile, Aztec, etc.