Remove `.hound.yml`
I run `grep swiftlint` in the context of replacing the CocoaPods
installation with a SwiftPM one,
https://github.com/wordpress-mobile/WordPress-iOS/pull/23958#discussion_r1920101295,
and got a hit on `.hound.yml`.

We have not been using Hound since at least a year, see paaHJt-68e-p2,
so there is no reason to keep carrying the config file in the repo.