Extract SFHFKeychainUtils in a dedicated Swift package
We need it in a dedicated package because the Objective-C implementation
does not use ARC. Therefore, we need to use the `-fno-objc-arc` compiler
setting. But such a setting should be applied sparingly so instead of
blanket using it in an existing Objective-C package like
WordPressSharedObjC, we use it surgically just for this file.