Adjust linking so that WordPress, Jetpack, and the tests build
The issue is that I named the framework and the package both
WordPressData and that we already have existing `import`s of it in
production and test code.

To keep the diff simple, I decided to leave those imports in place and
simply:

- Remove the WordPressData package (empty) dependency
- Add explicit WordPressDataObjc package (empty) dependency
- Add WordPressData framework (empty) dependency

This sorts the compilation and running locally.

As we'll move files to WordPressData and/or away from Keystone, we might
need to change this approach.