Fix #23025: Avoid unnecessary copying files
In the process of uploading an image from the Photos app, there are two
places that creates file using `mediaFileManager`, with the same
filename and at the same directory.
1. Copying the selected image file, which is in a temp directory, to the
   app's "local media directory". That's done by the code that's moved
   in this commit.
2. Creating a new image based on the selected image file, for the
   purpose of removing some exif data.

Since these two operations creates the same file at the same directory,
the second copy creates a file with "-1" suffix, which is uploaded to
the site's media library. That's the root cause of why there is a "-1"
in the title of uploaded images.