Clear the tag input field after tapping the Enter key (#25069)
We already have code that clears the text field in the `addTag`
function. However, during testing, I noticed it does not always work.
The new change here makes sure the `UITextField` is cleared, because we
are calling the UIKit directly, instead of relying on SwiftUI to pass
the `text` changes to UIKit.