Merge release/26.5.2 into trunk (#25071)
* Bump version number

* Fix CMM-1065: Find existing tag before creating a new one (#25061)

* Fix CMM-1065: Find existing tag before creating a new one

The issue was introduced in
https://github.com/wordpress-mobile/WordPress-iOS/pull/24955.

* Replace original input with the actual tag

* Invert a boolean check of most recently used tags first (#25065)

The issue was introduced in #24955. I missed a `!` in the refactor.

* Fix an issue where the app fails to remove tags from posts (#25067)

This is a regression introduced in #24964. We should not check if values
are an empty list in the "update" parameter types.

* Move finding existing tag code (#25068)

Moving the code to the view model, so that it can cover tags and custom
taxonomy terms, which are two seperate `TagsServiceProtocol`
implementations.

* Clear the tag input field after tapping the Enter key (#25069)

We already have code that clears the text field in the `addTag`
function. However, during testing, I noticed it does not always work.
The new change here makes sure the `UITextField` is cleared, because we
are calling the UIKit directly, instead of relying on SwiftUI to pass
the `text` changes to UIKit.

* Manually bump version number

---------

Co-authored-by: Tony Li <tony.li@automattic.com>