Remove `PostListFilter.oldestPostDate`
This property was added 9 years, and I don't think it applies to how the
posts are currently fetched and displayed anymore.

The current code fetched posts that is created after the `oldestPostDate`
(when set). However, the post list does not always sort by the "creation date".
Draft posts are sorted using modified date. So, the predicate created
based on `oldestPostDate` does not apply here.

There is also a related `fetchLimit` property. I'm not sure why we need
that. We should be displaying a full post list. There is no need to set
a limit here.